requirements.txt
aiohttp
pyserial